+++ docs/.vitepress/theme/index.ts
@@ -1,21 +1,21 @@
 // https://vitepress.dev/guide/custom-theme
-import { h } from 'vue'
-import Theme from 'vitepress/theme'
-import './style.css'
+import "./style.css";
+
+import { injectSpeedInsights } from "@vercel/speed-insights";
+import Theme from "vitepress/theme";
+import { h } from "vue";
+
+import Bazaar from "../components/bazaar.vue";
 
 export default {
     ...Theme,
     Layout: () => {
-        return h(Theme.Layout, null, {
-            // https://vitepress.dev/guide/extending-default-theme#layout-slots
-            'layout-bottom': h('script', {
-                'defer': '',
-                'src': '/_vercel/insights/script.js',
-            }),
-        });
+        return h(Theme.Layout);
     },
     enhanceApp(ctx) {
+        ctx.app.component("Bazaar", Bazaar);
+        injectSpeedInsights();
         Theme.enhanceApp(ctx);
         // ...
-    }
-}
+    },
+};